Separated LifeManager from Application Moved it to lifemanager.h and lifemanager.cpp Added lifemanager.cpp and lifemanager.h
