+++ application.h
@@ -17,7 +17,5 @@ private:
 
 	class Impl_;
 	std::unique_ptr<Impl_> impl_;
-	class LifeManager_;
-	std::unique_ptr<LifeManager_> lifeManager_;
 };
